few last changes in sendtomany tx

+++ src/tradelayer/tx.h
@@ -304,7 +304,7 @@ public:
     std::string getSpecial() const { return special; }
     std::string getPayload() const { return HexStr(pkt, pkt + pkt_size); }
     uint64_t getAmount() const { return nValue; }
-    uint64_t getAmountTotal() const { return std::accumulate(values.cbegin(), values.cend(), 0); }
+    uint64_t getAmountTotal() const { return std::accumulate(values.cbegin(), values.cend(), static_cast<uint64_t>(0)); }
     uint64_t getNewAmount() const { return nNewValue; }
     uint64_t getXAmount() const { return amount; }
     uint32_t getPreviousId() const { return prev_prop_id; }
